gst/tcp/gstmultifdsink.*: Implement stubbed out properties unit-type, units-soft-max, units-max, to allow specifying ...
Original commit message from CVS:
* gst/tcp/gstmultifdsink.c: (gst_sync_method_get_type),
(gst_multi_fd_sink_class_init), (get_buffers_max), (find_limits),
(gst_multi_fd_sink_recover_client),
(gst_multi_fd_sink_queue_buffer), (gst_multi_fd_sink_set_property),
(gst_multi_fd_sink_get_property):
* gst/tcp/gstmultifdsink.h:
Implement stubbed out properties unit-type, units-soft-max,
units-max, to allow specifying maximum sizes in units other than
buffers.
Fixes #355935

+++ gst/tcp/gstmultifdsink.h
@@ -213,8 +213,8 @@ struct _GstMultiFdSink {
   /* these values are used to check if a client is reading fast
    * enough and to control receovery */
   GstUnitType unit_type;/* the type of the units */
-  gint units_max;       /* max units to queue for a client */
-  gint units_soft_max;  /* max units a client can lag before recovery starts */
+  gint64 units_max;       /* max units to queue for a client */
+  gint64 units_soft_max;  /* max units a client can lag before recovery starts */
   GstRecoverPolicy recover_policy;
   GstClockTime timeout; /* max amount of nanoseconds to remain idle */
